Check all jobs are parseable Fixes #57
