+++ .gitignore
@@ -7,6 +7,8 @@ test/phantasma-flow/data
 #
 __debug_bin
 
+./phantasma-flow
+
 # Binaries for programs and plugins
 *.exe
 *.exe~
